# improve performance of sequenceT and sequenceS, fix #1255