**Before**

`sequenceT`

```
sequenceT (1) x 3,038,314 ops/sec ±1.06% (86 runs sampled)
sequenceT (2) x 1,486,717 ops/sec ±1.80% (86 runs sampled)
sequenceT (3) x 1,033,153 ops/sec ±0.79% (85 runs sampled)
sequenceT (4) x 799,994 ops/sec ±0.81% (87 runs sampled)
sequenceT (5) x 625,985 ops/sec ±0.80% (90 runs sampled)
```

`sequenceS`

```
sequenceS (1) x 2,723,960 ops/sec ±0.78% (86 runs sampled)
sequenceS (2) x 1,266,164 ops/sec ±0.74% (89 runs sampled)
sequenceS (3) x 809,389 ops/sec ±1.03% (89 runs sampled)
sequenceS (4) x 638,330 ops/sec ±1.43% (86 runs sampled)
sequenceS (5) x 522,986 ops/sec ±1.21% (89 runs sampled)
```

**After**

`sequenceT`

```
sequenceT (1) x 15,737,986 ops/sec ±0.71% (88 runs sampled)
sequenceT (2) x 8,389,840 ops/sec ±0.65% (86 runs sampled)
sequenceT (3) x 5,759,450 ops/sec ±0.54% (90 runs sampled)
sequenceT (4) x 4,307,637 ops/sec ±1.44% (85 runs sampled)
sequenceT (5) x 3,464,223 ops/sec ±1.75% (85 runs sampled)
```

`sequenceS`

```
sequenceS (1) x 13,783,868 ops/sec ±1.90% (82 runs sampled)
sequenceS (2) x 8,007,962 ops/sec ±1.55% (83 runs sampled)
sequenceS (3) x 4,969,669 ops/sec ±1.50% (86 runs sampled)
sequenceS (4) x 3,796,491 ops/sec ±0.98% (87 runs sampled)
sequenceS (5) x 2,973,628 ops/sec ±0.50% (89 runs sampled)
```

/cc @pbadenski